# backend/docs/openapi-schema.yml
openapi: 3.0.3
info:
  title: Ecommerce Web API
  version: 1.0.0
  description: REST API documentation for the Ecommerce Web platform, including authentication,
    catalog, cart, checkout, orders, payments, returns, inventory and administration.
  contact:
    name: Ecommerce Web Support
    email: support@example.com
paths:
...
  /api/tickets/{ticket_number}/reply/:
    post:
      operationId: tickets_reply_create
      parameters:
      - in: path
        name: ticket_number
        schema:
          type: string
        required: true
      tags:
      - tickets
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
  /api/tickets/management/tickets/{ticket_number}/assign/:
    patch:
      operationId: tickets_management_tickets_assign_partial_update
      parameters:
      - in: path
        name: ticket_number
        schema:
          type: string
        required: true
      tags:
      - tickets
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetail'
          description: ''
  /api/tickets/management/tickets/{ticket_number}/claim/:
    post:
      operationId: tickets_management_tickets_claim_create
      parameters:
      - in: path
        name: ticket_number
        schema:
          type: string
        required: true
      tags:
      - tickets
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetail'
          description: ''
  /api/tickets/management/tickets/{ticket_number}/reply/:
    post:
      operationId: tickets_management_tickets_reply_create
      parameters:
      - in: path
        name: ticket_number
        schema:
          type: string
        required: true
      tags:
      - tickets
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetail'
          description: ''
  /api/tickets/management/tickets/{ticket_number}/status/:
    patch:
      operationId: tickets_management_tickets_status_partial_update
      parameters:
      - in: path
        name: ticket_number
        schema:
          type: string
        required: true
      tags:
      - tickets
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetail'
          description: ''
...
components:
  schemas:
...
    CategoryEnum:
      enum:
      - ORDER_ISSUE
      - PAYMENT_ISSUE
      - DELIVERY_ISSUE
      - RETURN_REFUND
      - DAMAGED_PRODUCT
      - PRODUCT_INFORMATION
      - ACCOUNT_ISSUE
      - GENERAL_COMPLAINT
      type: string
      description: |-
        * `ORDER_ISSUE` - Order Issue
        * `PAYMENT_ISSUE` - Payment Issue
        * `DELIVERY_ISSUE` - Delivery Issue
        * `RETURN_REFUND` - Return or Refund
        * `DAMAGED_PRODUCT` - Damaged Product
        * `PRODUCT_INFORMATION` - Product Information
        * `ACCOUNT_ISSUE` - Account Issue
        * `GENERAL_COMPLAINT` - General Complaint
...
    PriorityEnum:
      enum:
      - LOW
      - MEDIUM
      - HIGH
      - URGENT
      type: string
      description: |-
        * `LOW` - Low
        * `MEDIUM` - Medium
        * `HIGH` - High
        * `URGENT` - Urgent
...
    RoleEnum:
      enum:
      - ADMIN
      - CUSTOMER
      - SUPPORT_AGENT
      - ORDER_MANAGER
      type: string
      description: |-
        * `ADMIN` - Admin
        * `CUSTOMER` - Customer
        * `SUPPORT_AGENT` - Support Agent
        * `ORDER_MANAGER` - Order Manager
...
    Status03eEnum:
      enum:
      - OPEN
      - ASSIGNED
      - IN_PROGRESS
      - WAITING_FOR_CUSTOMER
      - RESOLVED
      - CLOSED
      type: string
      description: |-
        * `OPEN` - Open
        * `ASSIGNED` - Assigned
        * `IN_PROGRESS` - In Progress
        * `WAITING_FOR_CUSTOMER` - Waiting for Customer
        * `RESOLVED` - Resolved
        * `CLOSED` - Closed
...
    TicketDetail:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        ticket_number:
          type: string
          readOnly: true
        customer:
          allOf:
          - $ref: '#/components/schemas/TicketUser'
          readOnly: true
        order_number:
          type: string
          readOnly: true
          nullable: true
        product:
          allOf:
          - $ref: '#/components/schemas/TicketProduct'
          readOnly: true
        assigned_agent:
          allOf:
          - $ref: '#/components/schemas/TicketUser'
          readOnly: true
        category:
          allOf:
          - $ref: '#/components/schemas/CategoryEnum'
          readOnly: true
        priority:
          allOf:
          - $ref: '#/components/schemas/PriorityEnum'
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/Status03eEnum'
          readOnly: true
        subject:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        resolved_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        closed_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        can_customer_reply:
          type: boolean
          readOnly: true
        can_customer_close:
          type: boolean
          readOnly: true
        messages:
          type: string
          readOnly: true
      required:
      - assigned_agent
      - can_customer_close
      - can_customer_reply
      - category
      - closed_at
      - created_at
      - customer
      - id
      - messages
      - order_number
      - priority
      - product
      - resolved_at
      - status
      - subject
      - ticket_number
      - updated_at
...
    TicketProduct:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          readOnly: true
        slug:
          type: string
          readOnly: true
          pattern: ^[-a-zA-Z0-9_]+$
        sku:
          type: string
          readOnly: true
      required:
      - id
      - name
      - sku
      - slug
    TicketUser:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          readOnly: true
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
        email:
          type: string
          format: email
          readOnly: true
        first_name:
          type: string
          readOnly: true
        last_name:
          type: string
          readOnly: true
        role:
          allOf:
          - $ref: '#/components/schemas/RoleEnum'
          readOnly: true
      required:
      - email
      - first_name
      - id
      - last_name
      - role
      - username
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT